Un-delay digital circuits
No need (that I can see, yet) to have "setTimeout" or other delays in here. Having them was complicating running the tests, because of the asynchronous nature. It is possible to have Jest tests wait until a Promise is done, but I was having a hard time figuring out how to return a correct promise.

For example, for an and circuit, we would really want calling "setWire" on one of the input wires to return a promise that would resolve when the entire circuit is done running, which would involve potentially many internal promises.